refactor: fix some ts issues

+++ src/api/bulk.ts
@@ -448,7 +448,7 @@ export class Job<
 
     this._jobInfo = (async () => {
       const jobId = await this.ready();
-      const body = ` 
+      const body = `
 <?xml version="1.0" encoding="UTF-8"?>
   <jobInfo xmlns="http://www.force.com/2009/06/asyncapi/dataload">
   <state>${state}</state>
@@ -1091,7 +1091,7 @@ export class BulkV2<S extends Schema> {
       await job.poll(options.pollInterval, options.pollTimeout);
       return await job.getAllResults();
     } catch (err) {
-      if (err.name !== 'JobPollingTimeoutError') {
+      if (err instanceof Error && err.name !== 'JobPollingTimeoutError') {
         // fires off one last attempt to clean up and ignores the result | error
         job.delete().catch((ignored) => ignored);
       }
@@ -1124,7 +1124,7 @@ export class BulkV2<S extends Schema> {
       await queryJob.poll(options?.pollInterval, options?.pollTimeout);
       return await queryJob.getResults();
     } catch (err) {
-      if (err.name !== 'JobPollingTimeoutError') {
+      if (err instanceof Error && err.name !== 'JobPollingTimeoutError') {
         // fires off one last attempt to clean up and ignores the result | error
         queryJob.delete().catch((ignored) => ignored);
       }
